requirements.txt
fastapi
uvicorn
google-generativeai>=0.8.3
pydantic
curl_cffi
requests
beautifulsoup4